+++ src/Form/FlickityForm.php
@@ -63,8 +63,7 @@ class FlickityForm extends EntityForm {
       '#type' => 'textfield',
       '#title' => $this->t('Cell selector'),
       '#default_value' => $options['cellSelector'],
-      '#description' => $this->t('Specify selector for cell elements. This is useful if you have other elements in your carousel elements that are not cells.'),
-      '#required' => TRUE
+      '#description' => $this->t('Specify selector for cell elements. This is useful if you have other elements in your carousel elements that are not cells.')
     );
 
     $form['setup']['initial_index'] = array(
@@ -73,6 +72,20 @@ class FlickityForm extends EntityForm {
       '#default_value' => $options['initialIndex'],
       '#min' => 0,
       '#description' => $this->t('Zero-based index of the initial selected cell.')
+    );
+
+    $form['setup']['set_gallery_size'] = array(
+      '#type' => 'checkbox',
+      '#title' => $this->t('Set gallery size'),
+      '#default_value' => $options['setGallerySize'],
+      '#description' => $this->t('Sets the height of the carousel to the height of the tallest cell. Set to false if you prefer to size the carousel with CSS, rather than using the size of cells.'),
+    );
+
+    $form['setup']['resize'] = array(
+      '#type' => 'checkbox',
+      '#title' => $this->t('Resize'),
+      '#default_value' => $options['resize'],
+      '#description' => $this->t('Adjusts sizes and positions when window is resized.'),
     );
 
     $form['cell_position'] = array(
@@ -92,6 +105,75 @@ class FlickityForm extends EntityForm {
         'right' => $this->t('Right')
       ),
       '#empty_value' => ''
+    );
+
+    $form['cell_position']['contain'] = array(
+      '#type' => 'checkbox',
+      '#title' => $this->t('Contain'),
+      '#default_value' => $options['contain'],
+      '#description' => $this->t('Contains cells to carousel element to prevent excess scroll at beginning or end. Has no effect if wrapAround is true.'),
+    );
+
+    $form['behavior'] = array(
+      '#type' => 'details',
+      '#title' => $this->t('Behavior'),
+      '#open' => TRUE
+    );
+
+    $form['behavior']['draggable'] = array(
+      '#type' => 'checkbox',
+      '#title' => $this->t('Draggable'),
+      '#default_value' => $options['draggable'],
+      '#description' => $this->t('Enable draggin and flicking.')
+    );
+
+    $form['behavior']['free_scroll'] = array(
+      '#type' => 'checkbox',
+      '#title' => $this->t('Free scroll'),
+      '#default_value' => $options['freeScroll'],
+      '#description' => $this->t('Enables content to be freely scrolled and flicked without aligning cells to an end position.')
+    );
+
+    $form['behavior']['wrap_around'] = array(
+      '#type' => 'checkbox',
+      '#title' => $this->t('Wrap around'),
+      '#default_value' => $options['wrapAround'],
+      '#description' => $this->t('At the end of cells, wrap-around to the other end for infinite scrolling.')
+    );
+
+    $form['behavior']['auto_play'] = array(
+      '#type' => 'number',
+      '#title' => $this->t('Auto play'),
+      '#default_value' => $options['autoPlay'],
+      '#min' => 0,
+      '#description' => $this->t('Milliseconds, set to 0 to disable. Automatically advances to the next cell. Auto-playing will pause when mouse is hovered over, and resume when mouse is hovered off. Auto-playing will stop when the carousel is clicked or a cell is selected.')
+    );
+
+    $form['behavior']['adaptive_height'] = array(
+      '#type' => 'checkbox',
+      '#title' => $this->t('Adaptive height'),
+      '#default_value' => $options['adaptiveHeight'],
+      '#description' => $this->t('Changes height of carousel to fit height of selected slide.')
+    );
+
+    $form['user_interface'] = array(
+      '#type' => 'details',
+      '#title' => $this->t('User interface'),
+      '#open' => TRUE
+    );
+
+    $form['user_interface']['prev_next_buttons'] = array(
+      '#type' => 'checkbox',
+      '#title' => $this->t('Prev/Next buttons'),
+      '#default_value' => $options['prevNextButtons'],
+      '#description' => $this->t('Creates and enables previous & next buttons.')
+    );
+
+    $form['user_interface']['page_dots'] = array(
+      '#type' => 'checkbox',
+      '#title' => $this->t('Page dots'),
+      '#default_value' => $options['pageDots'],
+      '#description' => $this->t('Creates and enables page dots. ')
     );
 
     $form['advanced_options'] = array(
@@ -128,6 +210,15 @@ class FlickityForm extends EntityForm {
       'cellSelector' => $form_state->getValue('cell_selector'),
       'initialIndex' => $form_state->getValue('initial_index'),
       'cellAlign' => $form_state->getValue('cell_align'),
+      'resize' => $form_state->getValue('resize'),
+      'setGallerySize' => $form_state->getValue('set_gallery_size'),
+      'contain' => $form_state->getValue('contain'),
+      'draggable' => $form_state->getValue('draggable'),
+      'freeScroll' => $form_state->getValue('free_scroll'),
+      'wrapAround' => $form_state->getValue('wrap_around'),
+      'autoPlay' => $form_state->getValue('auto_play'),
+      'prevNextButtons' => $form_state->getValue('prev_next_buttons'),
+      'pageDots' => $form_state->getValue('page_dots'),
       'advanced' => $form_state->getValue('advanced')
     ));