+++ src/Plugin/Field/FieldFormatter/EntityReferenceFlickityFormatter.php
@@ -91,10 +91,23 @@ class EntityReferenceFlickityFormatter extends EntityReferenceEntityFormatter {
    * @return array
    *   A renderable array for a themed field with its label and all its values.
    */
-  public function view(FieldItemListInterface $items, $langcode = NULL) {
+  public function viewElements(FieldItemListInterface $items, $langcode) {
 
-    $elements = parent::view($items, $langcode);
+    $elements = parent::viewElements($items, $langcode);
+    $config = $this->getSetting('flickity_config');
+
     $elements['#theme'] = 'flickity';
+    // Add Flickity data attribute.
+    $elements['#attributes']['data-flickity-options'] = $config;
+    // Attach library.
+    $elements['#attached']['library'][] = 'flickity/flickity';
+
+    // If not set, add flickity options.
+    if (!isset($elements['#attached']['drupalSettings']['flickity'][$config]) && !empty($config)) {
+      /** @var Flickity $entity */
+      $entity = Flickity::load($config);
+      $elements['#attached']['drupalSettings']['flickity'][$config] = $entity->getFormattedOptions();
+    }
 
     return $elements;
   }