+++ CHANGELOG.md
@@ -7,3 +7,6 @@ All notable changes to this project will be documented in this file.
 - Downloaded Flickity v2.0.5, added libraries.yml file. Flickity config entity.
 
 - Advanced options in Yaml format. Flickity field formatter.
+
+- Fixed getFormattedOptions(). Attached library to field. Implemented 
+flickityInit. Improved options form.